pyproject.toml
[tool.poetry]
name = "RMS"
version = "0.1.0"
description = "made as prize system for LTUC/ASAC"
authors = ["3madov-77 <emadzxy7@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.8"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.10.0"
django-filter = "^21.1"
django-debug-toolbar = "^3.2.2"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
psycopg2-binary = "^2.9.1"
django-environ = "^0.8.1"
coverage = "^6.1"
djangorestframework-simplejwt = "^5.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
